+++ internal/signalfx-agent/pkg/monitors/zcodegen/codegen.go
@@ -1,5 +1,4 @@
 //go:build !windows
-// +build !windows
 
 // Package codegen exists purely to manipulate the go:generate build order so
 // that collectd templates are generated first before the monitor metadata